Address review feedback: drop line refs, type temperature as float
Per @rlundeen2's review on PR #1861:

1. Replace `References:` blocks that cited line ranges in
   `gcg_attack.py` / `attack_manager.py` with symbol-only references.
   Line numbers drift the moment the legacy attack code is touched (B3
   wiring will do exactly that); symbol names are stable across the
   refactors that follow.

2. Re-type `SamplingStrategy.sample_candidates(temperature=)` as
   `float` instead of `int`. The protocol is a brand-new surface and
   was previously mirroring the legacy `GCGAlgorithmConfig.temp: int = 1`
   field for no good reason — sampling temperatures are conceptually
   continuous. The legacy field stays as-is; B3 wiring owns deciding
   whether to widen it or coerce at the boundary.

The stub used by the runtime-checkable tests is updated to match, and the
shape-smoke test now passes `temperature=1.0`.

+++ tests/unit/auxiliary_attacks/gcg/test_extension_protocols.py
@@ -73,7 +73,7 @@ class _StubSamplingStrategy:
         control_tokens: torch.Tensor,
         batch_size: int,
         top_k: int,
-        temperature: int,
+        temperature: float,
         allow_non_ascii: bool,
         non_ascii_tokens: torch.Tensor,
     ) -> torch.Tensor:
@@ -155,7 +155,7 @@ def test_sampling_strategy_stub_returns_expected_shape() -> None:
         control_tokens=control_tokens,
         batch_size=5,
         top_k=8,
-        temperature=1,
+        temperature=1.0,
         allow_non_ascii=True,
         non_ascii_tokens=non_ascii_tokens,
     )
